More reverts, cleanup of code, concurrent flag

+++ src/include/execution/exec/execution_settings.h
@@ -56,7 +56,7 @@ class EXPORT ExecutionSettings {
   bool GetIsCountersEnabled() const { return is_counters_enabled_; }
 
   /** @return number of threads used for parallel execution. */
-  size_t GetNumberofThreads() const { return number_of_threads_; }
+  int GetNumberofThreads() const { return number_of_threads_; }
 
   /** @return True if static partitioner is enabled. */
   constexpr bool GetIsStaticPartitionerEnabled() const { return is_static_partitioner_enabled_; }
